feat(lib): mark all nondet functions with noalias
Nondet functions are known not to acess any program memory.
Yet, the optimizer does not know that, which prevents some optimizations.

The only suitable attribute that I found in clang is __declspec(noalias).
This is not ideal. There are better attributes in LLVM IR, but I don't know
how to attach them at the source level.

If this will start causing problems, we'll have to solve it by an LLVM pass
and a naming convention for nondet functions.