fix(jobs): a different work around pointer analysis problem
Instead of running less precise pointer analysis, mark in the harness that two
pointers can alias. This lets the user tune the level of imprecision that is
desired, as opposed to completely giving up on type-awareness.

For this example, this makes no difference since there are very few pointers
used. However, I like that fix better since it is more visible to the user.

There is a plan to fix the pointer analysis as well, and then this annotation
will not matter.