feat(script): supress spurious vacuity errors
Don't report vacuity error if there is a pass message for the same line.
This is needed because an sassert can be split because of jumpthreading.
This leads to one BB being reachable and the other one being unreachable.
In this case vacuity actually passed even though it is unreachable in the other instance.