feat: config.h file for all explicit bounds
Currently, we have bounds being defined and re-defined in multiple places. This
makes it difficult to ensure that the right static bound is always used.

This commit adds a `include/config.h` file that is generate by cmake and has the
default values for all static constants. Code that depends on static constants
(such as `MAX_BUFFER_SIZE`) is expected to include the file via

```
```

The definitions in `config.h` are protected to allow overriding them from
command line. Thus, it is still possible to add custom overrides to these
parameters if desired.

The code has two ways to access static constants. Either using functions from
`bounds` library that are resolved at link time, or constants defined in
`config.h`. The use of `bounds` library is preferred, but sometimes a compile
time constant is necessary. In this case, `config.h` should be used.